requirements.txt
google-generativeai
wandb
trl
transformers
datasets
liger-kernel
# galore-torch
python-dotenv
# deepspeed
# accelerate
